add dtype

+++ tracklib/filter/dmmf.py
@@ -27,7 +27,7 @@ class IMMFilter(FilterBase):
         self._models = [model_cls[i](*init_args[i], **init_kwargs[i]) for i in range(self._models_n)]
         self._types = model_types
         if model_probs is None:
-            self._probs = np.full(self._models_n, 1 / self._models_n)
+            self._probs = np.full(self._models_n, 1 / self._models_n, dtype=float)
         else:
             self._probs = model_probs
         if trans_mat is None:
