+++ tracklib/filter/ukf.py
@@ -469,7 +469,7 @@ class SymmetricSigmaPoints():
 
     def init(self, dim):
         self._dim = dim
-        self._w = np.full(2 * dim, 1 / (2 * dim))
+        self._w = np.full(2 * dim, 1 / (2 * dim), dtype=float)
         self._init = True
 
     def points_num(self):
@@ -525,7 +525,7 @@ class ScaledSigmaPoints():
         if self._kappa is None:
             self._kappa = 3 - dim
         self._lamb = self._alpha**2 * (dim + self._kappa) - dim
-        self._w_mean = np.full(2 * dim + 1, 1 / (2 * (dim + self._lamb)))
+        self._w_mean = np.full(2 * dim + 1, 1 / (2 * (dim + self._lamb)), dtype=float)
         self._w_mean[-1] = self._lamb / (dim + self._lamb)
         self._w_cov = self._w_mean.copy()
         self._w_cov[-1] = self._w_mean[-1] + (1 - self._alpha**2 + self._beta)
